backend/requirements.txt
asgiref==3.8.1
asttokens==2.4.1
bcrypt==4.1.3
boto3==1.34.119
botocore==1.34.119
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==42.0.7
decorator==5.1.1
dj-rest-auth==6.0.0
Django==5.0.3
django-allauth==0.63.2
django-cors-headers==4.3.1
django-rest-swagger==2.2.0
django-soft-delete==1.0.13
django-storages==1.14.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
executing==2.0.1
idna==3.7
inflection==0.5.1
ipython==8.22.2
itypes==1.2.0
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mysqlclient==2.2.4
openapi-codec==1.3.2
packaging==24.0
paramiko==3.4.0
parso==0.8.3
pillow==10.3.0
prompt-toolkit==3.0.43
pure-eval==0.2.2
pycparser==2.22
Pygments==2.17.2
PyJWT==2.8.0
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
python-magic==0.4.27
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
s3transfer==0.10.1
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.14.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
wcwidth==0.2.13
